**** HTML plugin first implementation.
Raised a problem: plugin tagging is done on object creation.
But when MKNOD is called, object content is not in place, so
how could a plugin use the content to extract tags?
The answer is:

1) add a last_processed column in SQL objects table [done]
2) spawn a thread to periodically check old objects and process them [todo!]

**** IMPORTANT:

insert an object-type check when calling a processing plugin.
Only regular files or symlinks should be processed (not dirs,
pipes, devs, ...)

M    src/sql.c
M    src/tp_html.c
M    src/tagsistant.c